requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
Django==4.2.16
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
mysqlclient==2.2.4
numpy==2.0.2
packaging==24.1
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.0.3
